+++ src/domain/auth/components/birthdateForm/BirthdateForm.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Form, Formik, FormikProps } from 'formik';
 import { Button } from 'hds-react';
 import * as yup from 'yup';
@@ -9,6 +9,7 @@ import ageConstants from '../../../youthProfile/constants/ageConstants';
 import DateInput from '../../../../common/components/dateInput/DateInput';
 import styles from './BirthdateForm.module.css';
 
+const INVALID_DATE_ERROR = 'registration.dateInvalid';
 const INVALID_BIRTHDAY_ERROR = 'registration.birthdayInvalid';
 const AGE_RESTRICTION_ERROR = 'registration.ageRestriction';
 
@@ -30,19 +31,42 @@ type Props = {
 };
 
 function BirthdateForm(props: Props) {
+  // Most of the validation is done with a yup schema, but the existence
+  // of the inputted date is handled by the DateInput itself. As long as
+  // the date is invalid, the DateInput won't communicate it forward.
+
+  // This means that if the user inputs and invalid date (31.02.2020),
+  // the DateInput component will communicate that there is an invalid
+  // date error, while yup will communicate that there is a date does
+  // not exist error.
+
+  // Yup's error is checked later so it will override the error
+  // DateInput sets. Better UX would be to show the error about the
+  // invalid date. This is why we have a separate error state for for
+  // invalid date error. By using it, we are able to have more control
+  // over what error is shown when. I consider this sort of a hack
+  // because it's not very encapsulated. But for this case where we only
+  // have a single DateInput, it should be fine.
+  const [isDateInvalidError, setIsDateInvalid] = useState(false);
   const { t } = useTranslation();
 
-  const getError = (
+  const getDateError = (
     props: FormikProps<{ birthDate: string }>,
-    fieldName: string
+    fieldName = 'birthDate'
   ) => {
-    const { submitCount, errors } = props;
-    const isSubmitted = submitCount > 0;
+    const { errors, touched } = props;
     // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
     // @ts-ignore
-    const isError = Boolean(errors[fieldName]);
+    const isTouched = touched[fieldName];
+    // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+    // @ts-ignore
+    const isYupError = Boolean(errors[fieldName]);
 
-    if (isSubmitted && isError) {
+    if (isTouched && isDateInvalidError) {
+      return t(INVALID_DATE_ERROR);
+    }
+
+    if (isTouched && isYupError) {
       // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
       // @ts-ignore
       return t(errors[fieldName]);
@@ -69,10 +93,21 @@ function BirthdateForm(props: Props) {
             }
             onChange={date => {
               props.setFieldValue('birthDate', date.toJSON());
+              // If the date is passed on it should be valid so we can
+              // clear the error.
+              setIsDateInvalid(false);
             }}
-            error={getError(props, 'birthDate')}
+            onError={error => {
+              if (error.name === 'dateDoesNotExist') {
+                setIsDateInvalid(true);
+              }
+            }}
+            onBlur={() => {
+              props.setFieldTouched('birthDate', true);
+            }}
+            error={getDateError(props)}
             label={t('registration.birthdayHelper')}
-            dateInputLabel={'date'}
+            dateInputLabel="date"
             monthInputLabel="month"
             yearInputLabel="year"
             // Sets name in a way which allows auto-fill to set these